Adds asSet() to Bag for #726
This is limited to returning immutable or unmodifiable sets.
