+++ eclipse-collections-api/src/main/java/org/eclipse/collections/api/bag/MutableBag.java
@@ -228,4 +228,7 @@ public interface MutableBag<T>
      */
     @Override
     ImmutableBag<T> toImmutable();
+
+    @Override
+    MutableSet<T> asSet();
 }
