+++ eclipse-collections-api/src/main/java/org/eclipse/collections/api/bag/MutableBagIterable.java
@@ -127,4 +127,7 @@ public interface MutableBagIterable<T> extends Bag<T>, MutableCollection<T>
 
     @Override
     <V> RichIterable<V> collectWithOccurrences(ObjectIntToObjectFunction<? super T, ? extends V> function);
+
+    @Override
+    MutableSetIterable<T> asSet();
 }
