+++ eclipse-collections-api/src/main/java/org/eclipse/collections/api/bag/sorted/ImmutableSortedBag.java
@@ -257,4 +257,7 @@ public interface ImmutableSortedBag<T>
 
     @Override
     ImmutableSortedBag<T> drop(int count);
+
+    @Override
+    ImmutableSortedSet<T> asSet();
 }
