+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/immutable/ImmutableArrayBag.java
@@ -482,6 +482,12 @@ public class ImmutableArrayBag<T>
         return new ImmutableBagSerializationProxy<>(this);
     }
 
+    @Override
+    public ImmutableSet<T> asSet()
+    {
+        return new ImmutableSetFromArrayBagAdapter<>(this.keys);
+    }
+
     private final class ArrayBagIterator
             implements Iterator<T>
     {
