+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/mutable/MultiReaderHashBag.java
@@ -62,11 +62,13 @@ import org.eclipse.collections.api.map.MutableMap;
 import org.eclipse.collections.api.multimap.bag.MutableBagMultimap;
 import org.eclipse.collections.api.ordered.OrderedIterable;
 import org.eclipse.collections.api.partition.bag.PartitionMutableBag;
+import org.eclipse.collections.api.set.MultiReaderSet;
 import org.eclipse.collections.api.set.MutableSet;
 import org.eclipse.collections.api.tuple.Pair;
 import org.eclipse.collections.api.tuple.primitive.ObjectIntPair;
 import org.eclipse.collections.impl.collection.mutable.AbstractMultiReaderMutableCollection;
 import org.eclipse.collections.impl.factory.Iterables;
+import org.eclipse.collections.impl.set.mutable.MultiReaderSetAdapter;
 import org.eclipse.collections.impl.utility.LazyIterate;
 
 /**
@@ -639,6 +641,15 @@ public final class MultiReaderHashBag<T>
         }
     }
 
+    @Override
+    public MultiReaderSet<T> asSet()
+    {
+        try (LockWrapper wrapper = this.lockWrapper.acquireReadLock())
+        {
+            return MultiReaderSetAdapter.adapt(this.getDelegate().asSet(), this.lock);
+        }
+    }
+
     //Exposed for testing
 
     static final class UntouchableMutableBag<T>
@@ -694,6 +705,12 @@ public final class MultiReaderHashBag<T>
 
         @Override
         public MutableBag<T> asUnmodifiable()
+        {
+            throw new UnsupportedOperationException("cannot wrap an UntouchableMutableBag");
+        }
+
+        @Override
+        public MutableSet<T> asSet()
         {
             throw new UnsupportedOperationException("cannot wrap an UntouchableMutableBag");
         }
