+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/mutable/UnmodifiableBag.java
@@ -53,6 +53,7 @@ import org.eclipse.collections.api.set.MutableSet;
 import org.eclipse.collections.api.tuple.Pair;
 import org.eclipse.collections.api.tuple.primitive.ObjectIntPair;
 import org.eclipse.collections.impl.collection.mutable.AbstractUnmodifiableMutableCollection;
+import org.eclipse.collections.impl.set.mutable.UnmodifiableMutableSet;
 
 /**
  * An unmodifiable view of a bag.
@@ -437,5 +438,11 @@ public class UnmodifiableBag<T>
     public MutableSet<T> selectUnique()
     {
         return this.getMutableBag().selectUnique();
+    }
+
+    @Override
+    public MutableSet<T> asSet()
+    {
+        return UnmodifiableMutableSet.of(this.getMutableBag().asSet());
     }
 }
