+++ eclipse-collections/src/main/java/org/eclipse/collections/impl/bag/sorted/immutable/ImmutableEmptySortedBag.java
@@ -482,4 +482,10 @@ class ImmutableEmptySortedBag<T>
     {
         return SortedSets.immutable.of(this.comparator());
     }
+
+    @Override
+    public ImmutableSortedSet<T> asSet()
+    {
+        return SortedSets.immutable.of(this.comparator());
+    }
 }
